moved all locations filter dropdowns in a new package
