+++ mobile-app/app/src/main/java/com/semirsuljevic/raiffaisenmobileapp/ui/screens/intro_navbar/locations_filter/dropdowns/BranchServiceTypeDropdown.kt
@@ -1,4 +1,4 @@
-package com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.locations_filter
+package com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.locations_filter.dropdowns
 
 
 import androidx.compose.foundation.background
